% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{graph.family}
\alias{graph.family}
\title{A function that graphs a PCA run on a user-defined taxonomic family}
\usage{
graph.family(data, family.name)
}
\arguments{
\item{data}{dataframe including numerical data as well as information on taxonomic family and genus}

\item{family.name}{family name as it appears in the dataframe}
}
\description{
This is a sister function to pca.family() allowing you to visualize the analysis
}
\examples{
graph.family(data = d, family.name = "Cercopithecidae")
}
\keyword{PCA,}
\keyword{graph}
\keyword{taxonomic,}
