% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{graph.superfamily}
\alias{graph.superfamily}
\title{A function that graphs a PCA run on a user-defined taxonomic superfamily}
\usage{
graph.superfamily(data, superfamily.name)
}
\arguments{
\item{data}{dataframe including numerical data as well as information on taxonomic superfamily and family}

\item{superfamily.name}{superfamily name as it appears in the dataframe}
}
\description{
This is a sister function to pca.superfamily() allowing you to visualize the analysis
}
\examples{
graph.superfamily(data = d, superfamily.name = "Ceboidea")
}
\keyword{PCA,}
\keyword{graph}
\keyword{taxonomic,}
