% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{pca.family}
\alias{pca.family}
\title{A function that runs a PCA on a user-defined taxonomic family}
\usage{
pca.family(data, family.name)
}
\arguments{
\item{data}{dataframe including numerical data as well as information on taxonomic family and genus}

\item{family.name}{family name as it appears in the dataframe}
}
\description{
This function allows you to compare numerical data within a  taxonomic family
}
\examples{
pca.family(data = d, family.name = "Cercopithecidae")
}
\keyword{PCA,}
\keyword{taxonomic}
