% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{pca.superfamily}
\alias{pca.superfamily}
\title{A function that runs a PCA on a user-defined taxonomic superfamily}
\usage{
pca.superfamily(data, superfamily.name)
}
\arguments{
\item{data}{dataframe including numerical data as well as information on taxonomic superfamily and family}

\item{superfamily.name}{superfamily name as it appears in the dataframe}
}
\description{
This function allows you to compare numerical data within a  taxonomic superfamily
}
\examples{
pca.superfamily(data = d, superfamily.name = "Cebioidea")
}
\keyword{PCA,}
\keyword{taxonomic}
