% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopca-package.R
\docType{package}
\name{phylopca-package}
\alias{phylopca}
\alias{phylopca-package}
\title{phylopca: Phylogenetic Principle Component Analysis}
\description{
This package allows easy analysis of biological data using PCA.
}
\author{
\strong{Maintainer}: Lydia Myers \email{lydia.myers@utexas.edu}

}
\keyword{internal}
